Fixed set-associativity being more than 255 bug.
Bug Description:
The set associativity of the cache is configurable and when it's
configured with a value greater than 255, it is taken as 0 by the cache
code.

Fix Description:
The set associativity is stored as a uint8_t which can take only values
in the range [0, 255]. Changed it to uint16_t which can take values in
the range [0, 65535].

+++ src/cache.h
@@ -95,7 +95,7 @@ typedef struct cache_generic__ {
     char                name[CACHE_NAME_LEN];   /* name - L1, L2..          */
     char                trace_file[CACHE_TRACE_FILE_LEN];
     uint8_t             level;                  /* 1, 2, 3 ..               */
-    uint8_t             set_assoc;              /* level of associativity   */
+    uint16_t            set_assoc;              /* level of associativity   */
     uint32_t            blk_size;               /* cache block size         */
     uint32_t            size;                   /* total cache size         */
     uint8_t             repl_plcy;              /* replacement policy       */
